+++ packages/hds/jest.config.ts
@@ -1,4 +1,4 @@
-import { Config } from 'jest';
+import type { Config } from 'jest';
 
 const config: Config = {
   displayName: 'hds',
