+++ packages/hds/src/components/index.ts
@@ -1,5 +1,3 @@
 export * from './BreadcrumbSeparator';
-export * from './Fonts';
 export * from './GenericAvatarIcon';
-export * from './TimeInput';
-export * from './Toast/Toast';
+export * from './Toast';
