+++ packages/hds/src/theme/components/badge.ts
@@ -1,4 +1,4 @@
-import { StyleConfig } from '@chakra-ui/react';
+import type { StyleConfig } from '@chakra-ui/react';
 
 export const Badge: StyleConfig = {
   baseStyle() {
