+++ packages/hds/src/theme/components/checkbox.ts
@@ -1,4 +1,5 @@
-import { MultiStyleConfig, cssVar } from '@chakra-ui/react';
+import type { MultiStyleConfig} from '@chakra-ui/react';
+import { cssVar } from '@chakra-ui/react';
 import { createObjectFromKeysWithSameProps } from './_utils';
 
 const $shadow = cssVar('checkbox-shadow');
