+++ packages/hds/src/theme/components/close-button.ts
@@ -1,3 +1,3 @@
-import { StyleConfig } from '@chakra-ui/react';
+import type { StyleConfig } from '@chakra-ui/react';
 
 export const CloseButton: StyleConfig = {};
