+++ packages/hds/src/theme/components/form-error.ts
@@ -1,4 +1,4 @@
-import { MultiStyleConfig } from '@chakra-ui/react';
+import type { MultiStyleConfig } from '@chakra-ui/react';
 
 export const FormError: MultiStyleConfig = {
   parts: ['text', 'icon'],
