+++ packages/hds/src/theme/components/form-label.ts
@@ -1,4 +1,4 @@
-import { StyleConfig } from '@chakra-ui/react';
+import type { StyleConfig } from '@chakra-ui/react';
 
 export const FormLabel: StyleConfig = {
   baseStyle: {
