+++ packages/hds/src/theme/components/menu.ts
@@ -1,4 +1,5 @@
-import { MultiStyleConfig, cssVar } from '@chakra-ui/react';
+import type { MultiStyleConfig} from '@chakra-ui/react';
+import { cssVar } from '@chakra-ui/react';
 import { applyOpacity } from './_utils';
 
 const $shadow = cssVar('menu-shadow');
