+++ packages/hds/src/theme/components/modal.ts
@@ -1,4 +1,4 @@
-import { MultiStyleConfig } from '@chakra-ui/react';
+import type { MultiStyleConfig } from '@chakra-ui/react';
 
 export const Modal: MultiStyleConfig = {
   parts: [
