+++ packages/hds/src/theme/components/spinner.ts
@@ -1,4 +1,4 @@
-import { StyleConfig } from '@chakra-ui/react';
+import type { StyleConfig } from '@chakra-ui/react';
 import { isGrayish } from './_utils';
 
 export const Spinner: StyleConfig = {
