+++ packages/hds/src/theme/components/tabs.ts
@@ -1,4 +1,4 @@
-import { MultiStyleConfig } from '@chakra-ui/react';
+import type { MultiStyleConfig } from '@chakra-ui/react';
 
 export const Tabs: MultiStyleConfig = {
   parts: ['root', 'tab', 'tablist', 'tabpanels', 'tabpanel'],
