+++ packages/hds/src/theme/components/tag.ts
@@ -1,4 +1,4 @@
-import { MultiStyleConfig } from '@chakra-ui/react';
+import type { MultiStyleConfig } from '@chakra-ui/react';
 
 export const Tag: MultiStyleConfig = {
   parts: ['container', 'label', 'closeButton'],
