+++ packages/hds/src/theme/components/text.ts
@@ -1,4 +1,4 @@
-import { StyleConfig } from '@chakra-ui/react';
+import type { StyleConfig } from '@chakra-ui/react';
 
 const labels = {
   'label-xl': {
