--- packages/hds/src/theme/fontSizes.ts
@@ -1,6 +1,0 @@
-export const fontSizes = {
-  '2xs': '11px',
-  '3xs': '10px',
-  '4xs': '9px',
-  '5xs': '8px',
-};
