+++ packages/hds/src/theme/index.ts
@@ -3,20 +3,16 @@ import { breakpoints } from './breakpoints';
 import { colors } from './colors';
 import { components } from './components';
 import { config } from './config';
-import { fontSizes } from './fontSizes';
 import { fonts } from './fonts';
 import { styles } from './styles';
 import { textStyles } from './textStyles';
-import { zIndices } from './zIndices';
 
 export const theme = extendTheme({
-  config,
-  styles,
-  colors,
   breakpoints,
-  fonts,
+  colors,
   components,
-  fontSizes,
+  config,
+  fonts,
+  styles,
   textStyles,
-  zIndices,
 });
