+++ packages/hds/src/theme/styles.ts
@@ -1,4 +1,4 @@
-import { Styles } from '@chakra-ui/theme-tools';
+import type { Styles } from '@chakra-ui/theme-tools';
 
 export const styles: Styles = {
   global: {
