--- packages/hds/src/theme/zIndices.ts
@@ -1,1 +1,0 @@
-export const zIndices = {};
